Added functions, renamed functions, new explicit comments, moved betterft.h

+++ src/ft_itoa_un.c
@@ -6,13 +6,13 @@
 /*   By: rude-jes <ruipaulo.unify@outlook.fr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/20 20:03:30 by rude-jes          #+#    #+#             */
-/*   Updated: 2023/11/09 16:45:42 by rude-jes         ###   ########.fr       */
+/*   Updated: 2024/01/05 03:19:03 by rude-jes         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "betterft.h"
+#include "../betterft.h"
 
-char	*ft_itoa_un(unsigned int n)
+char	*ft_itoa_un(unsigned long n)
 {
 	char			*p;
 	unsigned int	temp;
